move module add views to core core dashboard
