+++ meinberlin/apps/dashboard/urls.py
@@ -21,13 +21,6 @@ urlpatterns = [
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/containers/$',
         ContainerListView.as_view(),
         name='container-list'),
-    url(r'^projects/(?P<project_slug>[-\w_]+)/blueprints/$',
-        views.ModuleBlueprintListView.as_view(),
-        name='module-blueprint-list'),
-    url(r'^projects/(?P<project_slug>[-\w_]+)/blueprints/'
-        '(?P<blueprint_slug>[-\w_]+)/$',
-        views.ModuleCreateView.as_view(),
-        name='module-create'),
     url(r'^organisations/(?P<organisation_slug>[-\w_]+)/plans/$',
         views.PlanListView.as_view(),
         name='plan-list'),
