+++ meinberlin/apps/dashboard2/templates/meinberlin_dashboard2/includes/module_blueprint_list_tile.html
