+++ meinberlin/apps/dashboard2/templates/meinberlin_dashboard2/module_blueprint_list.html
@@ -9,7 +9,7 @@
 
     <ul class="l-tiles-3">
         {% for blueprint_slug, blueprint in view.blueprints %}
-            {% include "meinberlin_dashboard/includes/module_blueprint_list_tile.html" with blueprint=blueprint %}
+            {% include "meinberlin_dashboard2/includes/module_blueprint_list_tile.html" with blueprint=blueprint %}
         {% endfor %}
     </ul>
 </div>
